+++ packages/colony-js-client/src/getNetworkClient.js
@@ -9,37 +9,7 @@ import NetworkLoader from '@colony/colony-js-contract-loader-network';
 import { TrufflepigLoader } from '@colony/colony-js-contract-loader-http';
 import ColonyNetworkClient from './ColonyNetworkClient/index';
 import EthersWrappedWallet from './lib/EthersWrappedWallet/index';
-
-export const defaultInfuraProjectId = '7d0d81d0919f4f05b9ab6634be01ee73';
-
-// Adapted from Ethers 4.0 to include support for Goerli
-const getInfuraProvider = (network: string, infuraProjectId?: string) => {
-  let host;
-  switch (network) {
-    case 'homestead':
-    case 'mainnet':
-      host = 'mainnet.infura.io';
-      break;
-    case 'goerli':
-      host = 'goerli.infura.io';
-      break;
-    default:
-      throw new Error(
-        `Could not get provider, unsupported network: ${network}`,
-      );
-  }
-
-  return new providers.JsonRpcProvider(
-    `https://${host}/v3/${infuraProjectId || defaultInfuraProjectId}`,
-    network === 'goerli'
-      ? {
-          chainId: 5,
-          ensAddress: '0x112234455c3a32fd11230c42e7bccd4a84e02010',
-          name: 'goerli',
-        }
-      : network,
-  );
-};
+import InfuraProvider from './lib/InfuraProvider/index';
 
 // This method provides a simple way of getting an initialized network client
 // that uses NetworkLoader for the remote network and TrufflePigLoader for a
@@ -65,7 +35,7 @@ const getNetworkClient = async (
     provider = new providers.JsonRpcProvider();
   } else {
     loader = new NetworkLoader({ network });
-    provider = getInfuraProvider(network, infuraProjectId);
+    provider = new InfuraProvider(network, infuraProjectId, verbose);
   }
 
   // Use EthersWrappedWallet if purser wallet