+++ packages/colony-js-client/src/lib/InfuraProvider/types.js
@@ -1,0 +1,7 @@
+/* @flow */
+
+export type Params = {
+  data: string,
+  from: string,
+  to: string,
+};